requirements.txt
djangorestframework
Pillow
drf_yasg
requests
python-decouple==3.4
google-auth